+++ src/datasources/data-frame/module.ts
@@ -1,10 +1,10 @@
 import { DataSourcePlugin } from '@grafana/data';
 import { DataFrameDataSource } from './DataFrameDataSource';
-import { DataFrameQueryEditor } from './components/DataFrameQueryEditor';
-import { HttpConfigEditor } from 'core/components/HttpConfigEditor';
-import { DataFrameVariableQueryEditor } from './components/DataFrameVariableQueryEditor';
+import { DataFrameQueryEditorWrapper } from './components/DataFrameQueryEditorWrapper';
+import { DataFrameVariableQueryEditorWrapper } from './components/DataFrameVariableQueryEditorWrapper';
+import { DataFrameConfigEditor } from './components/DataFrameConfigEditor';
 
 export const plugin = new DataSourcePlugin(DataFrameDataSource)
-  .setConfigEditor(HttpConfigEditor)
-  .setQueryEditor(DataFrameQueryEditor)
-  .setVariableQueryEditor(DataFrameVariableQueryEditor);
+  .setConfigEditor(DataFrameConfigEditor)
+  .setQueryEditor(DataFrameQueryEditorWrapper)
+  .setVariableQueryEditor(DataFrameVariableQueryEditorWrapper);